tools/requirements.txt
aiohttp>=3.9
aiolimiter>=1.1
beautifulsoup4>=4.13
markdown-it-py>=3.0
mdurl>=0.1.2
rapidfuzz>=3.8
scikit-learn>=1.5
joblib>=1.4
typer[all]>=0.12
urllib3>=2.2
python-slugify>=8.0
tqdm>=4.66
pyyaml>=6.0
# Optional (comment out if you don’t want embeddings initially)
sentence-transformers>=3.0
# Dev tools
black>=24.8.0
ruff>=0.6.0
pre-commit>=3.8.0
PyGithub>=2.5
